Change the method info logic to always create a new methodinfo for jitcompilationstarted and jitinlining. This used to try and cache the method infos, but this behavior is no longer supported.
Also, make the instrumentation engine log an error if a methodinfo
already exists. There are a few bad program exceptions happening
in the wild where the method body doesn't match the target
method info. The only way I see this could happen is if
one leaked and was reused. Note that the new logic will overwrite
any leaked methodinfos in the collections.