Added dynamic code unit tests.
Note, the dynamic code tests are designed to match existing behavior of
the CLRInstrumentationEngine. Whether or not this is the correct behavior
is up for discussion.

There are two types of dynamic code tested here.

1) Reflection.Emit.
   a) modules generated do not have a valid base load address
   b) modules generated do not have names that match what was set during
      the emit.
   c) methods in modules cannot be instrumented.

2) Code DOM
	a) modules generated do have a valid base load address
	b) modules generated have names that match what was set during
	   compilation
	c) methods in dynamic modules can be instrumented

In essence, we currently assume that if a module has a valid base load address,
then methods in that module can be instrumented. Otherwise, they cannot.