Implemented ICorProfilerInfo9 (#41)
* Implemented ICorProfilerInfo9

This change requires some removal and rearrangement of
dependent header files. Most importantly, we can no longer
depend on the NETFX supplied headers for corprof.h, since
it only includes the interfaces available for desktop. So,
a new cpp props file is added that ensures that all header
files are included in the correct order, making sure that
headers copied from dotnet core are included first.

* Code review comments

* Updated versions