Enable querying of IL instrumentation data (#42)
* Enable querying of IL instrumentation data

Adds the ability to read IL/Instrumentation
maps and IL/Native maps after jit completes.
To do this, we cache the IL/Instrumentation
maps that are generated during instrumentation
time and store them with the module that
contains the instrumented method until afer
jit completes.

The stored IL/Instrumentation map can be queried
via IMethodJitInfo2.

The IL/Instrumentation maps are also stored so
that we can work around a bug in the CLR:

The CLR doesn't correctly cache IL maps
in rejit scenarios. This means that after the
initial rejit, IL/Instrumentation and
IL/Native maps will be incorrect for other than
the first jit. To get around this, we reset the
instrumentation map in ICorProfilerInfo, and force
a query on the IL/Native map to ensure that the
internal IL/Instrumentation and IL/Native
maps are up-to-date within the CLR.