Optimize addref/release
A lot of time is spent in addref/release logic during
looping over instruction graphs with calls to
GetNextInstruction(...) and GetPreviousInstruction(...).
Most of these are not needed, and they add a lot of overhead
due to the standard calling convension and input/output checks.

1) Improves the speed by adding several functions that
convert the Previous/Next instruction functions to constexpr
so that they don't add function call overhead in internal code.

2) Adds new APIs that return pointers to the next/previous
instructions without ref counting so that clients of
CLRIE can execute their own loops faster.