Add back-links to branches from targets.
Makes branch targets track what branches target them.
This reduces the overhead of retargeting branches after
an insert because it means that the entire graph does
not need to be searched for branches. Instead the
branch target holds its own list of branches that target
it, and that list can be iterated over instead. This reduces
the cost of InsertBeforeAndRetargetOffsets from O(G) to O(B)
where B is the number of branches that target the Instruction
being inserted before. The amortized cost should therefore be
near O(1) because the average instruction is not a branch
target, and most branch targets will have only 1 branch tareting
them. I have not done the math to be certain of this, though.

The list of branches targeting a single instruction is stored
as a CDataItem so that we don't have to add additional pointers
to instructions where it isn't necessary. The list is a list
of raw pointers because every pointer in the list should be
kept alive by the graph, and using CComPtrs would increase overhead
while iterating through the list.