Add instrumentation engine unit tests.
Adds the ability for unit tests to test the single return instrumentation
method. This is useful because it is a complex instrumentation method
that rearanges IL in multiple ways, and generates a complex output. It is
also public instrumentation funcitonality that we don't want to break.

1) Updates the Naggler profiler to recognize a <MakeSingleRet /> command.
2) Fixes issue in instrumentation engine in which we were mis-reporting the original IL during method dumps.
3) Adds tests for single returns, testing if, switch, and exceptions.

Also adds tests for special edge-case around inserting branch instructions. When using insert before, and retargetting branches, the branch target can sometimes cause an infinite loop. This tests to ensure that doesn't happen. Finally, some failures in instrumentation may cause the instrumented function to hang or loop forever, so I've added a timeout for the process.