Test that demonstrate adding an exception handler
Set of unit test that show how an instrumentation method can be used to
wrap a method being jitted in an exception handler.

I got errors when for some of the tests, as CInstructionGraph::Replace
doesn't call CBranchTargetInfo::RetargetBranches. This means if you
replace an instruction that is a branch target, the branches will point
to the beginning of the method. This seems like a bug so I added a call
to CBranchTargetInfo::RetargetBranches in CInstructionGraph::Replace.