Update ubuntu image to use newer CMake
Updates the ubuntu image to add the kitware repo so that
we can get a newer version of CMake.

Also updates the DockerLocalBuild.ps1 so that the local
Dockerfiles can be used to generate images for building Linux
rather than depending on images that are only available to Microsoft
build pipelines.