Update alpine image with new cmake
Updates alpine image to download, build and install new
version of cmake so that we can make use of googletest
for native tests.

Adds cgmanifest file for the new cmake because our component
detection automation can't detect components downloaded
from the web in docker images.

Adds jq to the images so that we can query the cgmanifest for
versions, and keep our image building and component detection
in sync.