Fix tests
There was a problem with the previous logging for Function IDs during method dumps. Fixing the problem caused the tests to fail. This change updates the function ids in the baselines to be "Volatile" because they change on each execution of test.

Also fixes a typo that caused method IL rendering to fail.