Enable safe string cleanup (#452)
* Enable safe string cleanup

Adds APIs to CLRIE to enable safe and easy string cleanup.

Most of our strings from CLRIE are returned as BSTRs allocated inside the CLRIE module. On Windows, these strings are allocated via SysAllocString, and cleaned up using SysFreeString, which is easy. On other platforms, though, the are allocated via malloc() and free() using BSTR symmantics (essentially: the length of the string is in the two bytes prior to the string pointer, which is also null-terminated). So, clients are required to call "free" with the correct pointer (i.e. *not* the pointer to the first character), and on the correct c runtime.

Generally speaking, the c rutimes *should* be compatable between malloc and free, regardless of what module you are calling them from, but we shouldn't be relying on that, and we shouldn't be relying on clients understanding BSTR symantics.

So, this PR adds a new API: IProfilerStringManager, which gives a simple way to get an interface for freeing strings safely. It also adds a C-style interface called  "ClriFreeString" that clients can get a reference to using "GetProcAddress" or "dlsym".

Both ways of freeing strings are demonstrated in the Nagler instrumentation engine.